refactor: determineMovingRule을 MoveStrategy로 위임
