+++ src/main/java/janggi/domain/movestrategy/CannonStrategy.java
@@ -1,9 +1,11 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.board.Position;
+import janggi.domain.piece.Piece;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 public class CannonStrategy implements MoveStrategy {
 
@@ -22,5 +24,28 @@ public class CannonStrategy implements MoveStrategy {
         }
         path.add(to);
         return path;
+    }
+
+    @Override
+    public boolean determineMovingRule(Piece sourcePiece, Map<Position, Piece> positionPieces, Position to) {
+        if (positionPieces.size() >= 3) {
+            return false;
+        }
+        long count = positionPieces.values().stream()
+                .filter(piece -> sourcePiece.getPieceName().equals(piece.getPieceName()))
+                .count();
+        if (count != 0) {
+            return false;
+        }
+
+        if (positionPieces.size() == 2) {
+            if (positionPieces.containsKey(to)) {
+                Piece piece = positionPieces.get(to);
+                return !piece.isSameTeam(sourcePiece);
+            }
+            return false;
+        }
+
+        return !positionPieces.containsKey(to);
     }
 }
