+++ src/main/java/janggi/domain/movestrategy/ChariotStrategy.java
@@ -1,9 +1,11 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.board.Position;
+import janggi.domain.piece.Piece;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 public class ChariotStrategy implements MoveStrategy{
 
@@ -22,5 +24,19 @@ public class ChariotStrategy implements MoveStrategy{
         }
         path.add(to);
         return path;
+    }
+
+    @Override
+    public boolean determineMovingRule(Piece sourcePiece, Map<Position, Piece> positionPieces, Position to) {
+        if (positionPieces.size() >= 2) {
+            return false;
+        }
+        for (Position position : positionPieces.keySet()) {
+            if (position.equals(to)) {
+                return !sourcePiece.isSameTeam(positionPieces.get(position));
+            }
+            return false;
+        }
+        return true;
     }
 }
