+++ src/main/java/janggi/domain/movestrategy/GuardStrategy.java
@@ -1,8 +1,10 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.board.Position;
+import janggi.domain.piece.Piece;
 
 import java.util.List;
+import java.util.Map;
 
 public class GuardStrategy implements MoveStrategy {
 
@@ -15,4 +17,15 @@ public class GuardStrategy implements MoveStrategy {
     public List<Position> findPath(Position from, Position to) {
         return List.of(to);
     }
+
+    @Override
+    public boolean determineMovingRule(Piece sourcePiece, Map<Position, Piece> positionPieces, Position to) {
+        for (Piece piece : positionPieces.values()) {
+            if (piece.isSameTeam(sourcePiece)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
 }
