+++ src/main/java/janggi/domain/movestrategy/HorseStrategy.java
@@ -1,8 +1,10 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.board.Position;
+import janggi.domain.piece.Piece;
 
 import java.util.List;
+import java.util.Map;
 
 public class HorseStrategy implements MoveStrategy{
 
@@ -16,4 +18,19 @@ public class HorseStrategy implements MoveStrategy{
         Position mid = from.moveStraight(to);
         return List.of(mid, to);
     }
+
+    @Override
+    public boolean determineMovingRule(Piece sourcePiece, Map<Position, Piece> positionPieces, Position to) {
+        if (positionPieces.size() >= 2) {
+            return false;
+        }
+        for (Position position : positionPieces.keySet()) {
+            if (position.equals(to)) {
+                return !sourcePiece.isSameTeam(positionPieces.get(position));
+            }
+            return false;
+        }
+        return true;
+    }
+
 }
