+++ src/main/java/janggi/domain/movestrategy/SoliderStrategy.java
@@ -1,9 +1,11 @@
 package janggi.domain.movestrategy;
 
 import janggi.domain.board.Position;
+import janggi.domain.piece.Piece;
 import janggi.domain.piece.Team;
 
 import java.util.List;
+import java.util.Map;
 
 public class SoliderStrategy implements MoveStrategy {
     public static final int HAN_DIRECTION = 1;
@@ -31,4 +33,15 @@ public class SoliderStrategy implements MoveStrategy {
     public List<Position> findPath(Position from, Position to) {
         return List.of(to);
     }
+
+    @Override
+    public boolean determineMovingRule(Piece sourcePiece, Map<Position, Piece> positionPieces, Position to) {
+        for (Piece piece : positionPieces.values()) {
+            if (piece.isSameTeam(sourcePiece)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
 }
