+++ src/main/java/janggi/domain/piece/ChariotPiece.java
@@ -1,29 +1,10 @@
 package janggi.domain.piece;
 
-import janggi.domain.board.Position;
 import janggi.domain.movestrategy.MoveStrategy;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 
 public class ChariotPiece extends Piece {
     public ChariotPiece(Team team, MoveStrategy moveStrategy) {
         super(team, Name.CHARIOT, moveStrategy);
     }
 
-
-
-    @Override
-    public boolean determineMovingRule(Map<Position, Piece> positionPieces, Position to) {
-        if (positionPieces.size() >= 2) {
-            return false;
-        }
-        for (Position position : positionPieces.keySet()) {
-            if (position.equals(to)) {
-                return !isSameTeam(positionPieces.get(position));
-            }
-            return false;
-        }
-        return true;
-    }
 }
