+++ src/main/java/janggi/domain/piece/ElephantPiece.java
@@ -1,9 +1,6 @@
 package janggi.domain.piece;
 
-import janggi.domain.board.Position;
 import janggi.domain.movestrategy.MoveStrategy;
-
-import java.util.Map;
 
 public class ElephantPiece extends Piece {
 
@@ -11,17 +8,4 @@ public class ElephantPiece extends Piece {
         super(team, Name.ELEPHANT, moveStrategy);
     }
 
-    @Override
-    public boolean determineMovingRule(Map<Position, Piece> positionPieces, Position to) {
-        if (positionPieces.size() >= 2) {
-            return false;
-        }
-        for (Position position : positionPieces.keySet()) {
-            if (position.equals(to)) {
-                return !isSameTeam(positionPieces.get(position));
-            }
-            return false;
-        }
-        return true;
-    }
 }
