+++ src/main/java/janggi/domain/piece/GeneralPiece.java
@@ -1,9 +1,6 @@
 package janggi.domain.piece;
 
-import janggi.domain.board.Position;
 import janggi.domain.movestrategy.MoveStrategy;
-
-import java.util.Map;
 
 public class GeneralPiece extends Piece {
 
@@ -11,13 +8,4 @@ public class GeneralPiece extends Piece {
         super(team, Name.GENERAL, moveStrategy);
     }
 
-    @Override
-    public boolean determineMovingRule(Map<Position, Piece> positionPieces, Position to) {
-        for (Piece piece : positionPieces.values()) {
-            if (piece.isSameTeam(this)) {
-                return false;
-            }
-        }
-        return true;
-    }
 }
