+++ src/main/java/janggi/domain/piece/GuardPiece.java
@@ -1,9 +1,6 @@
 package janggi.domain.piece;
 
-import janggi.domain.board.Position;
 import janggi.domain.movestrategy.MoveStrategy;
-import java.util.List;
-import java.util.Map;
 
 public class GuardPiece extends Piece {
 
@@ -11,15 +8,4 @@ public class GuardPiece extends Piece {
         super(team, Name.GUARD, moveStrategy);
     }
 
-
-
-    @Override
-    public boolean determineMovingRule(Map<Position, Piece> positionPieces, Position to) {
-        for (Piece piece : positionPieces.values()) {
-            if (piece.isSameTeam(this)) {
-                return false;
-            }
-        }
-        return true;
-    }
 }
