+++ src/main/java/janggi/domain/piece/HorsePiece.java
@@ -1,10 +1,6 @@
 package janggi.domain.piece;
 
-import janggi.domain.board.Position;
 import janggi.domain.movestrategy.MoveStrategy;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 
 public class HorsePiece extends Piece {
 
@@ -12,19 +8,4 @@ public class HorsePiece extends Piece {
         super(team, Name.HORSE, moveStrategy);
     }
 
-
-
-    @Override
-    public boolean determineMovingRule(Map<Position, Piece> positionPieces, Position to) {
-        if (positionPieces.size() >= 2) {
-            return false;
-        }
-        for (Position position : positionPieces.keySet()) {
-            if (position.equals(to)) {
-                return !isSameTeam(positionPieces.get(position));
-            }
-            return false;
-        }
-        return true;
-    }
 }
